+++ backend/src/main/java/net/es/oscars/nsi/db/NsiMappingRepository.java
@@ -5,13 +5,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Repository
 public interface NsiMappingRepository extends JpaRepository<NsiMapping, Long> {
 
     List<NsiMapping> findAll();
-    List<NsiMapping> findByNsiConnectionId(String nsiConnectionId);
+    Optional<NsiMapping> findByNsiConnectionId(String nsiConnectionId);
     List<NsiMapping> findByNsiGri(String nsiGri);
     List<NsiMapping> findByOscarsConnectionId(String oscarsConnectionId);