+++ backend/src/main/java/net/es/oscars/resv/enums/DeploymentState.java
@@ -13,4 +13,9 @@ public enum DeploymentState {
     DEPLOY_FAILED,
 
     UNDEPLOY_FAILED,
+
+    WAITING_TO_BE_REDEPLOYED,
+    BEING_REDEPLOYED,
+    REDEPLOY_FAILED
+
 }