+++ backend/src/main/java/net/es/oscars/web/beans/ModifyResponse.java
@@ -3,16 +3,15 @@ package net.es.oscars.web.beans;
 import lombok.*;
 import net.es.oscars.resv.ent.Connection;
 
-import java.time.Instant;
-
 @Builder
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class ScheduleModifyResponse {
+public class ModifyResponse {
     @NonNull
     protected Boolean success;
-
     protected String explanation;
     protected Connection connection;
+
+
 }